Update error to preview error instead
