+++ code/lib/core-events/src/errors/server-errors.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import dedent from 'ts-dedent';
-import { UnknownArgTypesError, WebpackCompilationError } from './server-errors';
+import { WebpackCompilationError } from './server-errors';
 
 describe('WebpackCompilationError', () => {
   it('should correctly handle error with stats.compilation.errors', () => {
@@ -13,22 +12,5 @@ describe('WebpackCompilationError', () => {
 
     expect(webpackError.data.errors[0].message).toEqual('Error 1 message');
     expect(webpackError.data.errors[1].message).toEqual('Error 2 message');
-  });
-});
-
-describe('UnknownFlowArgTypesError', () => {
-  it('should correctly handle error with Flow convertSig', () => {
-    const type = {
-      name: 'signature',
-      type: 'number',
-      signature: 1,
-    };
-
-    const message = dedent`We detected a type {"name":"signature","type":"number","signature":1} in your configuration.
-    Your custom type does not match the TSFuncSigType or TSObjectSigType
-    Please check your Storybook configuration and ensure you have defined a valid type.`;
-
-    const typeError = new UnknownArgTypesError({ type });
-    expect(typeError.message).toEqual(message);
   });
 });
