+++ code/lib/docs-tools/src/argTypes/convert/flow/convert.ts
@@ -1,4 +1,4 @@
-import { UnknownArgTypesError } from '@storybook/core-events/server-errors';
+import { UnknownArgTypesError } from '@storybook/core-events/preview-errors';
 import type { SBType } from '@storybook/types';
 import type { FlowType, FlowSigType, FlowLiteralType } from './types';
 
@@ -19,7 +19,7 @@ const convertSig = (type: FlowSigType) => {
         value: values,
       };
     default:
-      throw new UnknownArgTypesError({ type: type });
+      throw new UnknownArgTypesError({ type: type, language: 'Flow' });
   }
 };
